refactor: Improve UI and navigation
Updated the welcome message in the Home screen from "Welcome," to "Good ".
Changed the navigation route for editing a project from a function call with the ID to a string route with a query parameter for the project ID.
Corrected the label and content description for the name input field in the Add/Edit screen from "Client Name" to "Project Name".
Updated the card title in the Project Detail screen from "Client Information" to "Project Information".
Replaced `Divider()` with `HorizontalDivider()` in the Project Detail screen.
Updated the style of the search bar label in the Search screen to `bodyMedium`.